Add components directory and Moved Components
