+++ client/src/components/Header.js
@@ -1,7 +1,7 @@
 import React, { Component } from "react";
 import { NavLink } from "react-router-dom";
 import socketIOClient from "socket.io-client";
-import "./header.css";
+import "./Header.css";
 
 // The Header creates links that can be used to navigate
 // between routes.
